refactor: move internal/infra under internal/provider/aws/infra (#627)
