+++ .golangci.yaml
@@ -97,7 +97,6 @@ linters:
         aws-sdk:
           files:
             - "!**/internal/api/**"
-            - "!**/internal/infra/**"
             - "!**/internal/provider/aws/**"
           deny:
             - pkg: "github.com/aws/aws-sdk-go-v2/service/ssm"
